api/requirements.txt
alembic==1.9.0
attrs==22.1.0
click==8.1.3
exceptiongroup==1.0.4
filetype==1.2.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
packaging==22.0
Pillow==9.3.0
pluggy==1.0.0
pytest==7.2.0
six==1.16.0
SQLAlchemy==1.4.45
tomli==2.0.1
Werkzeug==2.2.2
